% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceseq.R
\name{analyze_uridylation}
\alias{analyze_uridylation}
\title{Analyze uridylation}
\usage{
analyze_uridylation(dataset, transcript2, exp_type2,
  mapping_position_min = NA, mapping_position_max = NA, project = NA,
  facet_projects = FALSE, conditions = NA)
}
\arguments{
\item{dataset}{- dataset to analyze}

\item{transcript2}{- transcript to analyze}

\item{exp_type2}{- experiment type (OVR,KD,LEAP,NT)}

\item{mapping_position_min}{- min position of mapping in the reference}

\item{mapping_position_max}{- max position of mapping in the reference}

\item{project}{- project (project_name) to include in the analysis (character of character vector)}

\item{facet_projects}{- make facets based on projects (T/F)}

\item{conditions}{- filter for selected conditions (character vector)}
}
\value{
- list containing calculated uridylation data (calculated output), Dunn test results and lot
}
\description{
Analyze uridylation
}
\examples{
analyze_uridylation(all_data,"ACTB","OVR",project=c("Tailseq_4","Tailseq_5"),facet_projects==TRUE,conditions=c("CNTRL","TUT4WT","TUT7WT","MOV10"))
}
